requirements.txt
Pillow == 6.2.1
numpy == 1.19.5
torch == 1.6.0
torchvision == 0.7.0
